sceNpMatching2 stubs and some refactoring
