+++ rpcs3/emucore.vcxproj
@@ -597,6 +597,7 @@
     <ClInclude Include="Emu\SysCalls\Modules\libsnd3.h" />
     <ClInclude Include="Emu\SysCalls\Modules\libsynth2.h" />
     <ClInclude Include="Emu\SysCalls\Modules\sceNp.h" />
+    <ClInclude Include="Emu\SysCalls\Modules\sceNp2.h" />
     <ClInclude Include="Emu\SysCalls\Modules\sceNpClans.h" />
     <ClInclude Include="Emu\SysCalls\Modules\sceNpCommerce2.h" />
     <ClInclude Include="Emu\SysCalls\Modules\sceNpSns.h" />
