+++ rpcs3/emucore.vcxproj.filters
@@ -1825,5 +1825,8 @@
     <ClInclude Include="Emu\SysCalls\Modules\cellVideoOut.h">
       <Filter>Emu\SysCalls\Modules</Filter>
     </ClInclude>
+    <ClInclude Include="Emu\SysCalls\Modules\sceNp2.h">
+      <Filter>Emu\SysCalls\Modules</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
